refactor(nodedb-query): split expr.rs and scan_filter into focused modules
Move the monolithic `expr.rs` into an `expr/` directory (eval, codec,
binary, types submodules) so each concern lives in its own file.

Extract `FilterOp` and the `ScanFilter`/`CompareValue` types from
`scan_filter/mod.rs` into `op.rs` and `types.rs` respectively, leaving
`mod.rs` as pure re-exports. Add `like` and `parse` as explicit submodule
declarations now that the types they depend on are in separate files.

`msgpack_scan/filter.rs` updated to import from the new paths.
